snp_allele_counts; plot_snps
